hello_world/urls.py
"""hello_world URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.1/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from hello_world.core import views as core_views
from django.contrib import admin
from django.urls import path, include
import hello_world.settings.environments.development

urlpatterns = [
    path("", core_views.index),
    path("admin/", admin.site.urls),
    #path("__reload__/", include("django_browser_reload.urls")),
]

if hello_world.settings.environments.development.DEBUG:
    import debug_toolbar
    urlpatterns = [
        path('__debug__/', include(debug_toolbar.urls)),
    ] + urlpatterns
