pyproject.toml
[tool.poetry]
name = "codespaces-django"
version = "0.1.0"
description = ""
authors = ["Mikhail Tarbeev <tarbeev2003@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-debug-toolbar = "^4.2.0"
structlog = "^23.1.0"
autoconfig = "^5.0.0"
python-decouple = "^3.8"
django-browser-reload = "^1.11.0"
django-migration-linter = "^5.0.0"
django-test-migrations = "^1.3.0"
django-extra-checks = "^0.13.3"
psycopg2 = "^2.9.8"
django-split-settings = "^1.2.0"
django-stubs-ext = "^4.2.2"
django-cors-headers = "^4.2.0"
django-querycount = "^0.8.3"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
